Rename interfaces.py to schemas.py
Update imports in app.py and backend.py, and Dockerfile copy list.

Made-with: Cursor
